fix(ci): Resolve type checking errors and apply formatting fixes
- Enhanced decorator type signatures with TypeVar for better type preservation
- Fixed missing type arguments on Callable and dict types in adapters and tests
- Added comprehensive pyright configuration to handle decorator limitations
- Fixed tuple type annotation for access_history (int, float) not (int, int)
- Fixed unused variable issues in stress tests (replaced with underscore)
- Applied ruff formatting to all modified files
- Suppressed unavoidable type checking errors in decorator implementation

All tests pass (42/42), ruff checks pass, formatting applied.

+++ sdk/mycelium/protections/context_corruption.py
@@ -68,7 +68,7 @@ class ContextEntryHistory:
     source: str
     entity_id: str | None
     versions: list[ContextEntryVersion] = field(default_factory=list)
-    access_history: list[tuple[int, int]] = field(default_factory=list)  # (step, time)
+    access_history: list[tuple[int, float]] = field(default_factory=list)  # (step, time)
     invalidation_reasons: list[InvalidationReason] = field(default_factory=list)
 
     def current_version(self) -> ContextEntryVersion | None:
